Feat: Implement real-time path erasing
This commit introduces the functionality for users to erase paths in real-time during a collaborative session.

Key changes:
- Added `erasePaths` and `observeErase` to `RoomRepo` and its implementation `RoomRepoImp` to handle sending and receiving erase events via WebSockets.
- Updated `SocketManager` (Android and Desktop) to include `erase` and `observeErase` methods for WebSocket communication.
- Modified `PathRepo` and `PathRepoImp` to include `deleteAListOfPaths` for batch deletion of paths from the local database.
- Added `deleteListOfPaths` to `PathDao` to support deleting multiple paths by their IDs.
- In `WhiteBoardViewModel`:
    - Implemented `observeErase` to listen for incoming erase events and delete corresponding paths locally.
    - Added `erasePaths` function to send erase requests to the server when in Online mode.
    - Updated `upsertWhiteboard` to accept an `onSuccess` callback.
- Changed `opacity` type from `Int` to `Double` in `PathResponse.kt`.
- Minor adjustment in `RemoteDbApi` for fetching paths.

+++ composeApp/src/commonMain/kotlin/org/example/whiteboard/data/remote/dto/PathResponse.kt
@@ -8,7 +8,7 @@ data class PathResponse(
     val createdAt: String,
     val drawingTool: String,
     val fillColor: Int?,
-    val opacity: Int,
+    val opacity: Double,
     val pathId: String,
     val pathString: String,
     val strokeColor: Int,
